requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
